🧹 Extract file search logic to reusable findFiles helper
